добавлены маперы, репозитории, сервисы и их имплементации для чертежей и реализаций деталей. Начат процесс создания стандартных модулей и сохранения их в БД

+++ src/main/java/oleborn/robotsurvivalist/game/programlogics/model/entities/module/originalmodules/ProjectChassis.java
@@ -22,7 +22,7 @@ public class ProjectChassis extends ParentModule {
 
     //максимальная скорость
     @Column(name = "max_speed")
-    private Double maxSpeed;
+    private Integer maxSpeed;
 
     //возможность плавать
     @Column(name = "ability_to_float")
